Reorganize FeatureElements folder
